remove context pkg
