+++ decoder/candidates.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	icontext "github.com/hashicorp/hcl-lang/context"
 	"github.com/hashicorp/hcl-lang/lang"
 	"github.com/hashicorp/hcl-lang/schema"
 	"github.com/hashicorp/hcl/v2"
@@ -53,7 +52,7 @@ func (d *PathDecoder) candidatesAtPos(ctx context.Context, body *hclsyntax.Body,
 		if bodySchema.Extensions.Count {
 			if _, ok := body.Attributes["count"]; ok {
 				// append to context we need count completed
-				ctx = icontext.WithActiveCount(ctx)
+				ctx = schema.WithActiveCount(ctx)
 			}
 		}
 	}
